backend/borrowings-service/requirements.txt
django>=5.0,<5.2
djangorestframework
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.8
psycopg2-binary==2.9.11
django-cors-headers==4.6.0
django-environ==0.11.2
requests==2.32.3
gunicorn==23.0.0
django-filter